+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased][unreleased]
 
+- Add JavaScript interpretation
+
 ## [0.0.0][] - 2022-05-26
 
 - Create `CHANGELOG.md`